refactor: re-organiza arquivos relacionados com filtros
