+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/WithPagination/index.tsx
@@ -1,15 +1,19 @@
 'use client';
+import type React from "react";
 import type { GridPaginationModel } from "@mui/x-data-grid";
 import { type Dispatch, type SetStateAction, useState } from "react";
-import { DataTable } from "./modules/DataTable";
-import { PaginationContext } from "./PaginatedTable.context";
+import { PaginationContext } from "./context";
 
 export type PaginationModel = {
     gridPaginationModel: GridPaginationModel;
     onPaginationModelChange: Dispatch<SetStateAction<GridPaginationModel>>;
 }
 
-export const PaginatedTable = () => {
+type WithPaginationProps = React.PropsWithChildren
+
+export const WithPagination = ({
+    children,
+}: WithPaginationProps) => {
     const [paginationModel, setPaginationModel] = useState<GridPaginationModel>({
         page: 0,
         pageSize: 8,
@@ -17,7 +21,7 @@ export const PaginatedTable = () => {
     
     return (
         <PaginationContext.Provider value={{ gridPaginationModel: paginationModel, onPaginationModelChange: setPaginationModel }}>
-            <DataTable  />
+            {children}
         </PaginationContext.Provider>
     );
 };
